+++ src/codegen/gql2ts.ts
@@ -4,10 +4,16 @@ import {
   CodeGeneratorCliOptions,
 } from "./CodeGenerator";
 
-export async function gql2ts(
-  options: CodeGeneratorOptions &
-    CodeGeneratorCliOptions & { argv?: string[]; noExit?: boolean }
-) {
+export interface Gql2TsOptions
+  extends CodeGeneratorOptions,
+    CodeGeneratorCliOptions {
+  /**
+   * If true, does not exit the process on error.
+   */
+  noExit?: boolean;
+}
+
+export async function gql2ts(options: Gql2TsOptions) {
   const logger = options.logger ?? console;
 
   try {