+++ src/ws/GraphQLWebSocketClient.ts
@@ -6,13 +6,62 @@ import {
 import { AsyncPushIterator } from "../util/AsyncPushIterator";
 
 export interface GraphQLWebSocketClientOptions {
+  /**
+   * URL to GraphQL API.
+   */
   url: string;
+
+  /**
+   * WebSocket sub-protocol to use.
+   * Either "graphql-transport-ws" or "graphql-ws" (legacy).
+   * Defaults to "graphql-transport-ws".
+   */
   protocol?: string;
+
+  /**
+   * Payload to send during initialization.
+   * May be used server-side to acknowledge the connection.
+   */
   connectionInitPayload?: Record<string, unknown>;
+
+  /**
+   * Timeout to receive acknowledgement, in milliseconds.
+   * Defaults to 3000.
+   */
   connectionAckTimeout?: number;
+
+  /**
+   * Overrides the basic connection function.
+   * For example, when GraphQLWebSocketClient on server-side, you'd need
+   * to create a WebSocket from "ws" instead of a standard WebSocket.
+   */
   connect?: ConnectFn;
+
+  /**
+   * Should the client automatically reconnect on normal socket closure?
+   * Defaults to true.
+   */
+  autoReconnect?: boolean;
+
+  /**
+   * Minimum delay to wait before reconnection, in milliseconds.
+   * Reconnections only happen with open subscriptions and are controlled
+   * per subscription (see GraphQLWebSocketClient.subscribe()).
+   * The minimum is 10.
+   * Defaults to 500.
+   */
   minReconnectDelay?: number;
+
+  /**
+   * Maximum delay to wait before reconnection, in milliseconds.
+   * Defaults to 30000.
+   */
   maxReconnectDelay?: number;
+
+  /**
+   * Reconnection delay multiplier.
+   * Defaults to 2 (for exponential back-off).
+   */
   reconnectDelayMultiplier?: number;
 }
 
@@ -27,25 +76,21 @@ export interface SubscriptionRef {
   iterator: AsyncPushIterator<ExecutionResult>;
 }
 
-export interface SubscribeOptions {
-  maxRetries?: number;
-  minRetryDelay?: number;
-  maxRetryDelay?: number;
-  delayMultiplier?: number;
-}
-
 export class GraphQLWebSocketClient {
   public readonly url: string;
   public readonly protocol: string;
+  public gqlSocket?: GraphQLClientWebSocket;
   protected connectionInitPayload?: Record<string, unknown>;
   protected connectionAckTimeout: number;
   protected connectionAckPayload?: Record<string, unknown>;
   protected connect: ConnectFn;
+  protected autoReconnect: boolean;
   protected minReconnectDelay: number;
   protected maxReconnectDelay: number;
   protected reconnectDelayMultiplier: number;
   protected reconnectDelay: number;
-  protected socket?: GraphQLClientWebSocket;
+  protected reconnectDelayPromise?: Promise<void>;
+
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   protected subscriptions = new Set<AsyncPushIterator<any>>();
 
@@ -63,20 +108,18 @@ export class GraphQLWebSocketClient {
           connectionAckTimeout: this.connectionAckTimeout,
         }).init(connectionInitPayload));
 
-    this.minReconnectDelay = options.minReconnectDelay ?? 500;
+    this.autoReconnect = options.autoReconnect !== false;
+    this.minReconnectDelay = Math.max(options.minReconnectDelay ?? 500, 10);
     this.maxReconnectDelay = options.maxReconnectDelay ?? 30000;
     this.reconnectDelayMultiplier = options.reconnectDelayMultiplier ?? 2;
     this.reconnectDelay = 0;
   }
 
-  subscribe<TExecutionResult>(
-    payload: SubscribePayload,
-    options?: SubscribeOptions
-  ) {
+  subscribe<TExecutionResult>(payload: SubscribePayload) {
     return new AsyncPushIterator<TExecutionResult>(async (it) => {
       this.subscriptions.add(it);
 
-      const run = async (retries: number) => {
+      const run = async () => {
         try {
           const socket = await this.requireConnection();
           const results = socket.subscribe<TExecutionResult>(payload);
@@ -87,15 +130,15 @@ export class GraphQLWebSocketClient {
 
           it.finish();
         } catch (err) {
-          if (retries > 0 && this.shouldRetry(err)) {
-            run(retries - 1);
+          if (this.shouldRetry(err)) {
+            run();
           } else {
             it.throw(err);
           }
         }
       };
 
-      run(options?.maxRetries ?? 0);
+      run();
 
       return () => {
         this.subscriptions.delete(it);
@@ -108,7 +151,7 @@ export class GraphQLWebSocketClient {
       subscription.finish();
     }
 
-    this.socket?.close(code ?? 1000, reason ?? "Normal Closure");
+    this.gqlSocket?.close(code ?? 1000, reason ?? "Normal Closure");
   }
 
   shouldRetry(err: Error & { code?: number }) {
@@ -122,18 +165,14 @@ export class GraphQLWebSocketClient {
         return false;
     }
 
-    return true;
+    return typeof err.code === "number";
   }
 
   async requireConnection() {
-    if (!this.socket || !this.socket.isOpenOrConnecting()) {
-      if (this.reconnectDelay > 0) {
-        await new Promise((resolve) =>
-          setTimeout(resolve, this.reconnectDelay)
-        );
-      }
+    await this.maybeDelayReconnect();
 
-      this.socket = this.connect(
+    if (!this.gqlSocket || !this.gqlSocket.isOpenOrConnecting()) {
+      this.gqlSocket = this.connect(
         this.url,
         this.protocol,
         this.connectionInitPayload
@@ -141,21 +180,45 @@ export class GraphQLWebSocketClient {
     }
 
     try {
-      this.connectionAckPayload = await this.socket.requireAck();
-      this.reconnectDelay = this.minReconnectDelay;
+      this.connectionAckPayload = await this.gqlSocket.requireAck();
+      this.resetReconnectDelay();
 
-      return this.socket;
+      return this.gqlSocket;
     } catch (err) {
-      this.reconnectDelay = Math.floor(
-        Math.min(
-          this.reconnectDelayMultiplier *
-            Math.max(this.reconnectDelay, this.minReconnectDelay),
-          this.maxReconnectDelay
-        ) +
-          Math.random() * this.minReconnectDelay
-      );
+      this.increaseReconnectDelay();
 
       throw err;
     }
+  }
+
+  protected async maybeDelayReconnect() {
+    if (this.gqlSocket?.isOpenOrConnecting()) return;
+    if (this.reconnectDelay === 0) return;
+
+    if (!this.reconnectDelayPromise) {
+      this.reconnectDelayPromise = new Promise((resolve) => {
+        setTimeout(resolve, this.reconnectDelay);
+      });
+    }
+
+    if (this.reconnectDelayPromise) {
+      await this.reconnectDelayPromise;
+      this.reconnectDelayPromise = undefined;
+    }
+  }
+
+  protected resetReconnectDelay() {
+    this.reconnectDelay = this.minReconnectDelay;
+  }
+
+  protected increaseReconnectDelay() {
+    this.reconnectDelay = Math.floor(
+      Math.min(
+        this.reconnectDelayMultiplier *
+          Math.max(this.reconnectDelay, this.minReconnectDelay),
+        this.maxReconnectDelay
+      ) +
+        Math.random() * this.minReconnectDelay
+    );
   }
 }